refactor(server): extract OAuth callback routes to connectorRoutes.ts
Slice 2 of the server.ts split (follow-up to PR #97). Lifts the 7 connector
OAuth-callback routes (github/linear/sentry/google-calendar/google-drive/
slack/gmail) out of server.ts:589-721 and into a sibling export
`tryHandlePublicConnectorRoute` in src/connectorRoutes.ts.

Why a separate function rather than folding into tryHandleConnectorRoute:
the callback routes MUST run before the bearer-auth gate (they're
unauthenticated browser redirects from the OAuth vendor with no Patchwork
token), while the CRUD routes MUST run after it. Using one function with
two call sites would dispatch CRUD pre-auth on the first call, defeating
the gate.

Mechanical lift — no behavior change. Pre-extraction grep confirmed zero
`this.` references in the block. The Gmail callback's text/html content
type (vs. application/json for the others) is preserved.

server.ts shrinks 2495 → 2362 lines (-133, another -5%); cumulative across
both PRs the connector-related code is now -34% of the original.

Verified:
- Full bridge test suite: 322 files / 4492 tests green (no regressions)
- Zero LSP diagnostics on both files (errors + warnings)
- Biome clean
